+++ server/src-lib/Hasura/RQL/DDL/Metadata.hs
@@ -20,7 +20,9 @@ where
 import Control.Lens ((.~), (^.), (^?))
 import Data.Aeson qualified as J
 import Data.Aeson.Ordered qualified as AO
+import Data.Attoparsec.Text qualified as AT
 import Data.Bifunctor (bimap, first)
+import Data.Bitraversable
 import Data.ByteString.Lazy qualified as BL
 import Data.CaseInsensitive qualified as CI
 import Data.Environment qualified as Env
@@ -485,7 +487,14 @@ runTestWebhookTransform ::
   m EncJSON
 runTestWebhookTransform (TestWebhookTransform env urlE payload mt _ sv) = do
   url <- case urlE of
-    URL url' -> pure url'
+    URL url' ->
+      case AT.parseOnly parseEnvTemplate url' of
+        Left _ -> throwError $ err400 ParseFailed "Invalid Url Template"
+        Right xs ->
+          let lookup' var = maybe (Left var) (Right . T.pack) $ Env.lookupEnv env (T.unpack var)
+              result = traverse (fmap indistinct . bitraverse lookup' pure) xs
+              err e = throwError $ err400 NotFound $ "Missing Env Var: " <> e
+           in either err (pure . fold) result
     EnvVar var ->
       let err = throwError $ err400 NotFound "Missing Env Var"
        in maybe err (pure . T.pack) $ Env.lookupEnv env var
@@ -512,6 +521,15 @@ runTestWebhookTransform (TestWebhookTransform env urlE payload mt _ sv) = do
     -- NOTE: In the following two cases we have failed before producing a valid request.
     Left (UrlInterpError err) -> pure $ encJFromJValue $ J.toJSON err
     Left (RequestInitializationError err) -> pure $ encJFromJValue $ J.String $ "Error: " <> tshow err
+
+parseEnvTemplate :: AT.Parser [Either T.Text T.Text]
+parseEnvTemplate = AT.many1 $ pEnv <|> pLit <|> fmap Right "{"
+  where
+    pEnv = fmap (Left) $ "{{" *> AT.takeWhile1 (/= '}') <* "}}"
+    pLit = fmap Right $ AT.takeWhile1 (/= '{')
+
+indistinct :: Either a a -> a
+indistinct = either id id
 
 packTransformResult :: HTTP.Request -> J.Value -> EncJSON
 packTransformResult req body =